refactor: extract verification runtime
