+++ GUARD_CLASSIFICATION.md
@@ -15,15 +15,15 @@ not claim that every hard gate has already been moved to its target layer.
 | --- | --- | --- |
 | Canonical contract validation, digest and ID binding, later-turn approval, session and working-directory identity | `hooks/click_contract.py`; approval lifecycle in `hooks/click_gate.py` | Prevents mutation under an unapproved or substituted intent |
 | Mutation and external side-effect authorization | `hooks/click_mutation.py`; managed-service prepare/claim/record paths in `hooks/click_service.py` | Controls observable effects rather than model strategy |
-| Exact one-use runner token, action, request, root and state binding | mutation claims in `hooks/click_mutation.py`; managed-service claims in `hooks/click_service.py`; observation claims in `hooks/click_observation.py`; remaining verification claims in `hooks/click_gate.py` | Prevents replay, request substitution and cross-state execution |
+| Exact one-use runner token, action, request, root and state binding | mutation claims in `hooks/click_mutation.py`; managed-service claims in `hooks/click_service.py`; observation claims in `hooks/click_observation.py`; verification claims in `hooks/click_verification.py` | Prevents replay, request substitution and cross-state execution |
 | Canonical state paths, locks, atomic writes and authorized state-root recovery | `hooks/click_state.py` | Preserves approval state across failure without weakening identity checks |
 | Cancellation, expiry, consumed-runner and tampering revocation | `hooks/click_state.py`; lifecycle branches in `hooks/click_gate.py` | Prevents stale authority from becoming executable again |
-| Evidence registry identity, current-revision state and receipt matching | `hooks/click_evidence.py`; verification claim/result paths in `hooks/click_gate.py` | Prevents evidence from a different request or revision from completing work |
+| Evidence registry identity, current-revision state and receipt matching | `hooks/click_evidence.py`; verification claim/result paths in `hooks/click_verification.py` | Prevents evidence from a different request or revision from completing work |
 | Browser source admission, serial-call interlock, stable host-call/result mapping, current revision and finalization replay | `hooks/click_browser.py`; host routing in `hooks/click_gate.py` | Prevents an unassigned, parallel, mismatched, stale or replayed Browser result from completing evidence without judging the model's interaction strategy |
-| Protected Git-tree, environment and executable fingerprints for argv verification; verification-time mutation detection | verification receipt helpers and runners in `hooks/click_gate.py` | Preserves the identity of the code and environment actually checked by Click's argv runner |
+| Protected Git-tree, environment and executable fingerprints for argv verification; verification-time mutation detection | receipt helpers and runners in `hooks/click_verification.py` | Preserves the identity of the code and environment actually checked by Click's argv runner |
 | Opt-in dependency provider, relevant-entry and resolved-path receipts, plus approved mutation pre/post binding | `hooks/click_dependency_cache.py`; mutation-boundary and receipt transitions in `hooks/click_mutation.py` | Allows cross-revision reuse without treating a model's post-approval assertion, unrelated manifest content, or later workspace drift as proof |
 | Direct-argv capability safety, executable/path/environment checks, process-control rejection and read-only side-effect defenses | shared protocol validation in `hooks/click_capability.py`; inspection policy and hardened execution in `hooks/click_inspection.py`; mutation admission in `hooks/click_mutation.py` | Prevents an admitted inspect, verify or service request from gaining unapproved effects |
-| Mutation, observation, verification and managed-service interlocks around active runner claims | mutation reservation and claim state in `hooks/click_mutation.py`; observation reservations and claims in `hooks/click_observation.py`; managed-service state in `hooks/click_service.py`; remaining verification state in `hooks/click_gate.py` | Prevents an authorized side effect or receipt from racing into a different state without treating distinct observation concurrency as authority |
+| Mutation, observation, verification and managed-service interlocks around active runner claims | mutation reservation and claim state in `hooks/click_mutation.py`; observation reservations and claims in `hooks/click_observation.py`; verification state in `hooks/click_verification.py`; managed-service state in `hooks/click_service.py` | Prevents an authorized side effect or receipt from racing into a different state without treating distinct observation concurrency as authority |
 | Host coverage identity, event normalization and state/receipt-preserving result mapping | `hooks/click_host_coverage.py`, `hooks/click_hook.py`, `hooks/antigravity_gate.py`, `hooks/platform_protocol.py` | Keeps the known pre/post surface symmetric and prevents verification receipts from crossing a host or coverage-registry revision while making the limited assurance explicit |
 
 ## USER_POLICY
@@ -43,7 +43,7 @@ not claim that every hard gate has already been moved to its target layer.
 | Advising on `update_plan` or equivalent host plan tools | plan-tool branch in `hooks/click_gate.py`; host matchers and adapters | Advisory only — hard denial removed in v0.30.0 |
 | Choosing the proposed evidence source and qualitative verification profile before approval, concrete argv during execution, and the cheapest proof strategy | Skill, eval and documentation policy | Model strategy or an explicit user constraint; Click records the choice but does not interpret it as runtime authority |
 | Legacy verification class-unit values | `hooks/click_verification_policy.py` and `hooks/click_verification_meter.py` | Compatibility data only — they do not produce runtime advice, prove cost or sufficiency, or participate in receipt authority |
-| Requiring a mutation after a fixed number of otherwise legitimate failed retries | argv verification and Browser retry handling | Advisory only in v0.30.0; fresh authorization remains distinct from replay of an active or consumed runner |
+| Requiring a mutation after a fixed number of otherwise legitimate failed retries | argv handling in `hooks/click_verification.py` and Browser retry handling | Advisory only in v0.30.0; fresh authorization remains distinct from replay of an active or consumed runner |
 | Browser repeat/retry guidance, preferred timing thresholds and interaction-history depth | `hooks/click_browser_advisory.py`; bounded history compaction in Browser preparation | Advisory only — normalized repeats and long timed interactions remain receipt-bound and allowed; old per-input guidance is compacted instead of blocking a new call |
 | Contract compactness, planning prose and workflow-shape preferences beyond parser or transport safety | `hooks/click_contract.py` schema policy | Keep only identity-bearing protocol fields in Core |
 | Main Skill authoring compactness | reference split, plugin validation and review | Heuristic documentation quality — no word-count permission gate; required links and protocol structure remain testable without optimizing prose to an arbitrary number |
